+++ src/lib.rs
@@ -8,3 +8,5 @@ pub mod timeline;
 pub mod yaml;
 #[macro_use]
 extern crate prettytable;
+#[macro_use]
+extern crate horrorshow;